logarithmic distribution method and test case added.

+++ example/logarithmic_test.cc
@@ -10,12 +10,12 @@ using namespace std;
 
 int main()
 {
-    variate_generator<double, logarithmic> vg(0.33);    
+    variate_generator<int, logarithmic> vg(0.33);    
 
-    copy( vg.begin(), vg.begin()+1000, ostream_iterator<double>(cout, "\n"));
+    copy( vg.begin(), vg.begin()+1000, ostream_iterator<int>(cout, "\n"));
 
 
-    auto sum = accumulate( vg.begin(), vg.begin()+1000, 0.0 );
+    auto sum = accumulate( vg.begin(), vg.begin()+1000, 0 );
 
     cout << sum << endl;
